Fix collision detection!!
It's not 100% correct at all times as I would REALLY like it to be, but certainly
good enough for now, and maybe forever.
With low frame rates it can almost certainly be incorrect, as the ball could
travel through the block and come out on the other side, and therefore bounce as if
it came from that side.