Fix collision edge cases!
(Or literally corner cases? :-)

When a ball collides with a brick, we only checked the position relative to the brick.
As I expected prior to even implementing the brick collisions, that's not enough to always
give a correct result.

These changes check the ball velocity direction to figure out the correct collision direction
in such cases.
So far it has been correct 100% of the time when I have checked them with the debugger.
I'm not sure it DETECTS 100% of the errors (but I believe it does), but when it HAS detected
them, it has successfully corrected them.