Add GenericDialog and use for ConfirmationDialog
The idea was to also move LoadingScreen to use GenericDialog with null
for buttons, but it causes runtime issues, possibly due to circular
dependencies, though I'm not sure.

If that is the cause, I'm not sure why ConfirmationDialog works, as it
also has circular dependencies: WindowManager depends on
ConfirmationDialog, which depends on GenericDialog, which depends on
WindowManager. This causes no issues.